+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added authority claimer service to support reader mode
 - Added support to `POST` *inspect state* requests
 
+### Removed
+
+- Removed configurable inspect-server path prefix
+
 ## [1.0.1] 2023-09-06
 
 ### Fixed